refactor: remove redudant functions from Shishido
These include functions like split_by_demog

The next step would be refactor unnecessary code to create the nmf
model from Shishido, which is needlessly complex. The only purpose
of including it is to ensure that the final 4 visualizations run

Also, will need to abstract the code for visualizations into one
function instead of repeating the code for each variable.