bug: fix the list flattening error in split_incorrect
The function returns lists for conjoined words and added them to a
an existing list of correctly spelt words. This created issues with
the final join function, which cannot operate on nested lists.

I fixed this by changing the return value of split word to a string
using the joun function.

Only one error remains with this cleaning setup- how to deal with the
edge case of misspelt subwords- such as 'thesimularities'. There may
be some recursive approach to find the most probable split.